+++ NEWS.md
@@ -5,6 +5,7 @@
 * Exported the `qenv` class from the package.
 * The `@code` field in the `qenv` class now holds `character`, not `expression`.
 * The `get_code` method  returns a single concatenated string of the code.
+* Removed internal `format_expression` function.
 
 # teal.code 0.4.1