+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.code 0.4.1.9009
+# teal.code 0.4.1.9010
 
 ### Miscellaneous